# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for pet adoption platform",
        "title": "Pet Adoption API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/api/pet/all": {
            "get": {
                "description": "Retrieve all pets",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pets"
                ],
                "summary": "Get all pets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.PetSearchFilterResponse"
                        }
                    }
                }
            }
        },
        "/api/pet/register": {
            "post": {
                "description": "Register a new pet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pets"
                ],
                "summary": "Register pet",
                "parameters": [
                    {
                        "description": "Pet Payload",
                        "name": "pet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.PetRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.PetRegisterResponse"
                        }
                    }
                }
            }
        },
        "/api/pet/{pid}": {
            "get": {
                "description": "Retrieve pet by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pets"
                ],
                "summary": "Get pet info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pet ID",
                        "name": "pid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.PetGetInfoByIdResponse"
                        }
                    }
                }
            }
        },
        "/api/user/delete": {
            "delete": {
                "description": "Delete authenticated user",
                "tags": [
                    "users"
                ],
                "summary": "Delete user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.UserRegisterResponse"
                        }
                    }
                }
            }
        },
        "/api/user/info": {
            "get": {
                "description": "Get authenticated user profile",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.UserDeleteResponse"
                        }
                    }
                }
            }
        },
        "/api/user/login": {
            "post": {
                "description": "Login using email and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "Login Payload",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UserLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.UserLoginResponse"
                        }
                    }
                }
            }
        },
        "/api/user/oauth/{provider}": {
            "get": {
                "description": "Redirect user to OAuth provider",
                "tags": [
                    "users"
                ],
                "summary": "Start OAuth login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth Provider",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/oauth/{provider}/callback": {
            "get": {
                "description": "Handle OAuth provider callback",
                "tags": [
                    "users"
                ],
                "summary": "OAuth callback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth Provider",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/register": {
            "post": {
                "description": "Create new user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Register user",
                "parameters": [
                    {
                        "description": "Register Payload",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UserRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.UserRegisterResponse"
                        }
                    }
                }
            }
        },
        "/api/user/update": {
            "put": {
                "description": "Update authenticated user's profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update user",
                "parameters": [
                    {
                        "description": "Update Payload",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UserUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.UserUpdateResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.PetSearchFilterResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "petsInfo": {}
            }
        },
        "domain.PetGetInfoByIdResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "petInfo": {}
            }
        },
        "domain.PetRegisterRequest": {
            "type": "object",
            "required": [
                "petName"
            ],
            "properties": {
                "petDetail": {
                    "type": "string"
                },
                "petName": {
                    "type": "string",
                    "minLength": 0
                }
            }
        },
        "domain.PetRegisterResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "pid": {
                    "type": "integer"
                }
            }
        },
        "domain.UserDeleteResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "domain.UserLoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "domain.UserLoginResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "domain.UserRegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "firstName",
                "lastName",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string",
                    "minLength": 1
                },
                "lastName": {
                    "type": "string",
                    "minLength": 1
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "domain.UserRegisterResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "domain.UserUpdateRequest": {
            "type": "object",
            "properties": {
                "firstName": {
                    "type": "string",
                    "minLength": 1
                },
                "lastName": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
        "domain.UserUpdateResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}